requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.28.1
google-generativeai==0.3.2
